the python file that do the functions
FileReading.py is for the wild type TEV. It creates the overlap and non-overlap data of the Naive and Sorted library in a CSV file, by updating dictionaries with the compassion with the (Naive&Sorted) list. However, all the sequences are in one row. The table was transposed and reconstructed in Excel later. 

FileReading_PE10 is for PE10 TEV. It creates similar data like WildType TEV but using Dataframe from pandas. The non-overlap data of sorted and naive are summarized clearly in sheets "NonOverlap_Sorted" and "NonOverlap_Naive" respectively in the excel file: "PEScatterData.xlsx". The overlapping data was summarized in the sheet named"Overlap"